% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{animate}
\alias{animate}
\title{Automatic Creation of Animations}
\usage{
animate(object, ...)
}
\arguments{
\item{object}{the input object.}

\item{...}{currently of no use; included for future compatibility.}
}
\value{
This generic does not return anything and is just the basis for more 
sophisticated methods.
}
\description{
A generic that is the basis for methods that allow the user to create 
specific animations on-the-fly.
}
\details{
A generic that can be extended by methods to automatically create animations 
based on certain objects.
}
