% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{CIVLABOR}
\alias{CIVLABOR}
\title{Monthly Civilian Labor Force Level in the USA}
\format{
An object of class \code{ts} of length 864.
}
\source{
The data was obtained from the databank of the Federal Reserve Bank of 
St. Louis (accessed: 2022-09-01) and then transformed into a time series
object using R.

\url{https://fred.stlouisfed.org/series/LNU01000000}
}
\usage{
CIVLABOR
}
\description{
A \code{ts} object that contains the monthly observed civilian labor force
level in the USA (in millions of persons) from January 1948 to December 2019. 
The object contains 864 observations.
}
\keyword{datasets}
