% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{CONSUMPTION}
\alias{CONSUMPTION}
\title{Quarterly Real Final Consumption Expenditure for Australia}
\format{
An object of class \code{ts} with 242 rows and 1 columns.
}
\source{
The data was obtained from the databank of the Federal Reserve Bank 
of St. Louis
(accessed: 2023-09-26) and 
then transformed into a time series
object using R. 

\url{https://fred.stlouisfed.org/series/NCRNSAXDCAUQ}
}
\usage{
CONSUMPTION
}
\description{
A \code{ts} object that contains the quarterly real final consumption 
expenditure (in thousands of domestic currency) from the third quarter of 1959 
to the last quarter in  2019. The object contains 242 observations.
}
\keyword{datasets}
