% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{COVID}
\alias{COVID}
\title{Daily Confirmed New COVID-19 Cases in Germany}
\format{
An object of class \code{ts} of length 183.
}
\source{
The data was obtained from the COVID-19 Data Hub 
(accessed: 2023-09-25) and 
then transformed into a time series
object using R. 

\url{https://covid19datahub.io/}
}
\usage{
COVID
}
\description{
A \code{ts} object that contains the daily confirmed new COVID-19 cases
in Germany (in thousands of cases) from June 2021 to November 2021.
The object contains 183 observations. The time series object is created 
as a time series object with frequency 7, i.e. the time unit is in calendar 
weeks of the year 2021 and not in years.
}
\keyword{datasets}
