% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{ENERGY}
\alias{ENERGY}
\title{Monthly Total Production and Distribution of Electricity, Gas, Steam, and Air Conditioning for Germany}
\format{
An object of class \code{ts} with 390 rows and 1 columns.
}
\source{
The data was obtained from the databank of the Federal Reserve Bank 
of St. Louis
(accessed: 2023-10-11) and 
then transformed into a time series
object using R. 

\url{https://fred.stlouisfed.org/series/DEUPREND401IXOBM}
}
\usage{
ENERGY
}
\description{
A \code{ts} object that contains the monthly production and distribution 
of electricity gas steam, and air conditioning for Germany (Index 2015 = 100) 
from January 1991 
to June 2023. The object contains 390 observations.
}
\keyword{datasets}
