% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{EXPENDITURES}
\alias{EXPENDITURES}
\title{Quarterly Personal Consumption Expenditures in the USA}
\format{
An object of class \code{ts} of length 292.
}
\source{
The data was obtained from the databank of the Federal Reserve Bank of 
St. Louis (accessed: 2022-07-15) and then transformed into a time series
object using R.

\url{https://fred.stlouisfed.org/series/NA000349Q}
}
\usage{
EXPENDITURES
}
\description{
A \code{ts} object that contains the quarterly observed personal
consumption expenditures in the USA in trillions of dollars from the
first quarter of 1947 to the last quarter of 2019. The object contains
292 observations.
}
\keyword{datasets}
