% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-deseats_fc.R
\name{expo,deseats_fc-method}
\alias{expo,deseats_fc-method}
\title{Exponentiate \code{deseats} Forecasts}
\usage{
\S4method{expo}{deseats_fc}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"deseats_fc"}.}

\item{...}{currently without purpose; included for compatibility only.}
}
\value{
An object of class \code{"deseats_fc"} is returned.
}
\description{
Exponentiate, i.e. as act of retransformation, (point and interval) forecasts 
obtained via the \code{deseats} package.
}
\details{
This function simply exponentiates point forecasts and bounds of interval
forecasts within an object of class \code{"deseats_fc"}. This does not
have the option to account for potential biases in point forecasts. For this
purpose, use the predict method directly with \code{expo = TRUE} and
\code{adjust.bias = TRUE}.
}
\examples{
\donttest{
est <- s_semiarma(log(EXPENDITURES), set_options(order_poly = 3))
fc <- predict(est, n.ahead = 8)
fc2 <- expo(fc)
fc2
}

}
