% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{expo}
\alias{expo}
\title{Automatic Creation of Animations}
\usage{
expo(object, ...)
}
\arguments{
\item{object}{the input object.}

\item{...}{currently of no use; included for future compatibility.}
}
\value{
This generic does not return anything and is just the basis for more 
sophisticated methods.
}
\description{
A generic that is the basis for methods that allow the user to exponentiate 
certain results quickly.
}
\details{
A generic that can be extended by methods to exponentiate the numeric 
information in certain objects.
}
