% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gain_function.R
\name{gain}
\alias{gain}
\title{Gain Function Generic}
\usage{
gain(object, lambda = seq(0, 0.5, 1e-04), ...)
}
\arguments{
\item{object}{an input object; for the linear filters considered in this 
object, gain function values should be obtained.}

\item{lambda}{a vector of frequencies (from 0 to 0.5) for which 
to obtain the gain function.}

\item{...}{currently without use; for possible future compatibility.}
}
\value{
This generic does not return anything and is just the basis for more 
sophisticated methods.
}
\description{
Gain Function Generic
}
\details{
A standard generic function. The purpose is to build various methods 
to instantaneously obtain gain function values for linear filters of 
different decomposition objects.

A generic that can be extended by methods to automatically obtain gain 
function values of linear filters described in certain objects.
}
\author{
\itemize{
\item Dominik Schulz (Research Assistant) (Department of Economics, Paderborn
University), \cr
Author and Package Creator
}
}
