% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{GDP}
\alias{GDP}
\title{Quarterly US GDP}
\format{
An object of class \code{ts} with 292 rows and 1 columns.
}
\source{
The data was obtained from the databank of the Federal Reserve Bank 
of St. Louis
(accessed: 2023-09-25) and 
then transformed into a time series
object using R. 

\url{https://fred.stlouisfed.org/series/NA000334Q}
}
\usage{
GDP
}
\description{
A \code{ts} object that contains the quarterly US GDP
(in billions of USD) from the first quarter of 1947 to the last quarter in 
2019. The object contains 292 observations.
}
\keyword{datasets}
