% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{HOUSES}
\alias{HOUSES}
\title{Monthly New One Family Houses Sold in the USA}
\format{
An object of class \code{ts} with 252 rows and 1 columns.
}
\source{
The data was obtained from the databank of the Federal Reserve Bank 
of St. Louis
(accessed: 2023-09-30) and 
then transformed into a time series
object using R. 

\url{https://fred.stlouisfed.org/series/HSN1FNSA}
}
\usage{
HOUSES
}
\description{
A \code{ts} object that contains the monthly new one family houses sold in 
the USA (in thousands of units) from January 1985 
to December 2005. The object contains 252 observations.
}
\keyword{datasets}
