% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{LIVEBIRTHS}
\alias{LIVEBIRTHS}
\title{Monthly Live Births in Germany}
\format{
An object of class \code{ts} of length 339.
}
\source{
The data was obtained from the databank "Genesis" of the National 
Statistical Office of Germany (accessed: 2023-06-15) and then transformed into 
a time series object using R.

\url{https://www-genesis.destatis.de/genesis/online?operation=previous&levelindex=4&levelid=1686843345946&levelid=1686843308009&step=3#abreadcrumb}
}
\usage{
LIVEBIRTHS
}
\description{
A \code{ts} object that contains the monthly observed live births in Germany
(in thousands of cases) from January 1995 to March 2023. The object 
contains 339 observations.
}
\keyword{datasets}
