% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{NOLABORFORCE}
\alias{NOLABORFORCE}
\title{Monthly Number of US Persons Not in the Labor Force}
\format{
An object of class \code{ts} of length 360.
}
\source{
The data was obtained from the databank of the Federal Reserve Bank of 
St. Louis (accessed: 2023-06-15) and then transformed into a time series
object using R.

\url{https://fred.stlouisfed.org/series/LNU05000000}
}
\usage{
NOLABORFORCE
}
\description{
A \code{ts} object that contains the monthly observed number of persons in
the USA that do not belong to the labor force (in millions of persons)
from January 1990 to December 2019. 
The object contains 360 observations.
}
\keyword{datasets}
