% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-smoothing_options.R
\name{order_poly,smoothing_options-method}
\alias{order_poly,smoothing_options-method}
\alias{order_poly<-,smoothing_options-method}
\alias{season,smoothing_options-method}
\alias{season<-,smoothing_options-method}
\alias{kernel_fun,smoothing_options-method}
\alias{kernel_fun<-,smoothing_options-method}
\alias{bwidth,smoothing_options-method}
\alias{bwidth<-,smoothing_options-method}
\alias{boundary_method,smoothing_options-method}
\alias{boundary_method<-,smoothing_options-method}
\title{Retrieve or Set Smoothing Options}
\usage{
\S4method{order_poly}{smoothing_options}(object)

\S4method{order_poly}{smoothing_options}(object) <- value

\S4method{season}{smoothing_options}(object)

\S4method{season}{smoothing_options}(object) <- value

\S4method{kernel_fun}{smoothing_options}(object)

\S4method{kernel_fun}{smoothing_options}(object) <- value

\S4method{bwidth}{smoothing_options}(object)

\S4method{bwidth}{smoothing_options}(object) <- value

\S4method{boundary_method}{smoothing_options}(object)

\S4method{boundary_method}{smoothing_options}(object) <- value
}
\arguments{
\item{object}{an object of class \code{"smoothing_options"}.}

\item{value}{the value to set the corresponding option to.}
}
\value{
The methods without \code{<-} return the corresponding settings in the supplied 
object. The methods with \code{<-} set the corresponding option in the 
provided object.
}
\description{
Retrieve smoothing options from or set them in an object 
of class \code{"smoothing_options"}.
}
\details{
Various methods are provided to either retrieve smoothing option
settings from an object of class \code{"smoothing_options"} or 
to adjust them within such an object.
}
\examples{
opts <- set_options()
opts
order_poly(opts)
order_poly(opts) <- 3
opts

}
\author{
\itemize{
\item Dominik Schulz (Research Assistant) (Department of Economics, Paderborn
University), \cr
Author and Package Creator
}
}
