% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{RETAIL}
\alias{RETAIL}
\title{Monthly Total Volume of Retail Trade in Germany}
\format{
An object of class \code{ts} with 348 rows and 1 columns.
}
\source{
The data was obtained from the databank of the Federal Reserve Bank 
of St. Louis
(accessed: 2023-10-11) and 
then transformed into a time series
object using R. 

\url{https://fred.stlouisfed.org/series/DEUSLRTTO01IXOBM}
}
\usage{
RETAIL
}
\description{
A \code{ts} object that contains the monthly total volume of retail trade 
in Germany (Index 2015 = 100)
from January 1991 
to December 2019. The object contains 348 observations.
}
\keyword{datasets}
