% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDecomposition.R
\name{runDecomposition}
\alias{runDecomposition}
\title{Shiny App for Decomposing Seasonal Time Series}
\usage{
runDecomposition()
}
\value{
This function returns \code{NULL}.
}
\description{
A shiny app is started that allows its user to run some of the decomposition
approaches of this package interactively.
}
\details{
Time series data can be uploaded to the app and then decomposed using 
some of the approaches implemented in this package. The decomposition 
is immediately visualized within the app, so that the user can assess the
suitability of the decomposition graphically. The decomposed data can then be 
saved in CSV format.
}
\author{
\itemize{
\item Dominik Schulz (Research Assistant) (Department of Economics, Paderborn
University), \cr
Author and Package Creator
}
}
