% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{SAVINGS}
\alias{SAVINGS}
\title{Quarterly Savings of Private Households in Germany}
\format{
An object of class \code{ts} of length 116.
}
\source{
The data was obtained from the databank "Genesis" of the National 
Statistical Office of Germany (accessed: 2023-06-15) and then transformed into 
a time series object using R.

\url{https://www-genesis.destatis.de/genesis/online?operation=abruftabelleBearbeiten&levelindex=1&levelid=1686857137066&auswahloperation=abruftabelleAuspraegungAuswaehlen&auswahlverzeichnis=ordnungsstruktur&auswahlziel=werteabruf&code=81000-0010&auswahltext=&werteabruf=Werteabruf#abreadcrumb}
}
\usage{
SAVINGS
}
\description{
A \code{ts} object that contains the quarterly observed savings of
private households in Germany
(in billions of EUR) from January 1991 to December 2019. The object 
contains 116 observations.
}
\keyword{datasets}
