% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonplot.R
\name{seasonplot_gg}
\alias{seasonplot_gg}
\title{Creation of Seasonal Plots in the Style of ggplot2}
\usage{
seasonplot_gg(x)
}
\arguments{
\item{x}{a time series object of class \code{"ts"}.}
}
\value{
The function returns a \code{ggplot2} plot object.
}
\description{
Simplified seasonal plot creation of time series objects in order to
identify seasonal patterns.
}
\details{
The function returns an plot object in the style of \code{ggplot2}. The plot
can therefore be adjusted easily using common \code{ggplot2} syntax.
}
\examples{
seasonplot_gg(TEMPERATURE) +
 ggplot2::ylab("Degrees Celsius") +
 ggplot2::ggtitle("Seasonal plot of German temperature data")

}
