% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-deseats.R
\name{show,deseats-method}
\alias{show,deseats-method}
\title{Printing of \code{deseats} Function Results}
\usage{
\S4method{show}{deseats}(object)
}
\arguments{
\item{object}{an object of class \code{"deseats"}.}
}
\value{
The function returns \code{NULL}.
}
\description{
Print results of the function \code{deseats}.
}
\details{
After trend and seasonality in a time series have been estimated using the
function \code{deseats}, basic estimation results can be easily printed to 
the console.
}
