% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-s_semiarma.R
\name{show,s_semiarma-method}
\alias{show,s_semiarma-method}
\title{Show Method for Objects of Class \code{"s_semiarma"}}
\usage{
\S4method{show}{s_semiarma}(object)
}
\arguments{
\item{object}{an object of class \code{"s_semiarma"}.}
}
\value{
This method returns \code{NULL}.
}
\description{
Print results of the function \code{\link{s_semiarma}} to the console.
}
\details{
Use this method to create a nice looking overview of the contents of 
objects of class \code{"s_semiarma"}.
}
\author{
\itemize{
\item Dominik Schulz (Research Assistant) (Department of Economics, Paderborn
University), \cr
Author and Package Creator
}
}
