% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{SUNSHINE}
\alias{SUNSHINE}
\title{Monthly Hours of Sunshine in Germany}
\format{
An object of class \code{ts} of length 872.
}
\source{
The data was obtained from the databank of the 
German Weather Service (DWD) (accessed: 2023-09-25) and 
then transformed into a time series
object using R. 

\url{https://www.dwd.de/DE/leistungen/zeitreihen/zeitreihen.html#buehneTop}
}
\usage{
SUNSHINE
}
\description{
A \code{ts} object that contains the monthly observed hours of sunshine
in Germany from January 1951 to August 2023.
The object contains 872 observations.
}
\keyword{datasets}
