% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{TEMPERATURE}
\alias{TEMPERATURE}
\title{Monthly Average Temperature in Germany}
\format{
An object of class \code{ts} of length 1712.
}
\source{
The data was obtained from the databank of the 
German Weather Service (DWD) (accessed: 2023-09-25) and 
then transformed into a time series
object using R. 

\url{https://www.dwd.de/DE/leistungen/zeitreihen/zeitreihen.html#buehneTop}
}
\usage{
TEMPERATURE
}
\description{
A \code{ts} object that contains the monthly observed average temperature
in Germany (in degrees Celsius) from January 1881 to August 2023.
The object contains 1712 observations.
}
\keyword{datasets}
