% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{trend}
\alias{trend}
\alias{deseasonalize}
\alias{detrend}
\title{Obtain Estimated Components of a Time Series}
\usage{
trend(object, ...)

deseasonalize(object, ...)

detrend(object, ...)
}
\arguments{
\item{object}{the estimation object.}

\item{...}{currently without use; included for future compatibility.}
}
\value{
These generics do not return anything and are just the basis for more 
sophisticated methods.
}
\description{
Obtain estimated components, such as the estimated trend, the seasonally 
adjusted series, and so on from an estimation object.
}
\details{
Generics that can be extended by methods to obtain (estimated) time series
components from certain estimation objects.
}
